refac: move package structure
