+++ performance/wazuh_manager_mockers/src/manager_mock_services/agent_comm_mock/middlewares/brotli.py
