+++ performance/wazuh_manager_mockers/src/manager_mock_services/manager_server_mock/manager_server_mock.py
@@ -54,11 +54,11 @@ import uvicorn
 from fastapi import Depends, FastAPI, Header, HTTPException
 from fastapi.responses import JSONResponse
 
-from manager_mock_servers.manager_services.manager_server_mock.models import AgentData, AuthData
-from manager_mock_servers.utils.token_manager import TokenManager
-from manager_mock_servers.utils.agent_database_handler import insert_new_agent, check_if_agent_exists, check_if_uuid_exists, \
+from manager_mock_services.manager_server_mock.models import AgentData, AuthData
+from utils.token_manager import TokenManager
+from utils.agent_database_handler import insert_new_agent, check_if_agent_exists, check_if_uuid_exists, \
     create_agents_database
-from manager_mock_servers.utils.vars import (
+from utils.vars import (
     DEFAULT_AUD,
     DEFAULT_EXPIRATION_TIME,
     DEFAULT_ISS,
