+++ performance/wazuh_manager_mockers/src/wazuh_manager_mockers.egg-info/PKG-INFO
@@ -1,3 +1,22 @@
+Metadata-Version: 2.1
+Name: wazuh_manager_mockers
+Version: 0.0.1
+Summary: Wazuh Library for mocking manager services
+Author-email: Wazuh <hello@wazuh.com>
+Classifier: Programming Language :: Python :: 3
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.10
+Description-Content-Type: text/markdown
+Requires-Dist: Brotli==1.1.0
+Requires-Dist: cryptography==43.0.0
+Requires-Dist: fastapi==0.112.0
+Requires-Dist: PyJWT==2.9.0
+Requires-Dist: pyproject_hooks==1.1.0
+Requires-Dist: uvicorn==0.30.5
+Provides-Extra: test
+Requires-Dist: pytest==8.3.2; extra == "test"
+Requires-Dist: httpx==0.27.0; extra == "test"
+
 ## Manager Service mocks
 
 This package provides scripts and tools to simulate the Agent Communication and Management Server API services of the Wazuh manager. These mocks are specifically designed for performance testing of the agent and are not intended for use in a production environment.
