+++ performance/wazuh_manager_mockers/src/wazuh_manager_mockers.egg-info/requires.txt
@@ -1,0 +1,10 @@
+Brotli==1.1.0
+cryptography==43.0.0
+fastapi==0.112.0
+PyJWT==2.9.0
+pyproject_hooks==1.1.0
+uvicorn==0.30.5
+
+[test]
+pytest==8.3.2
+httpx==0.27.0
