+++ performance/wazuh_manager_mockers/src/wazuh_manager_mockers.egg-info/top_level.txt
@@ -1,0 +1,4 @@
+__init__
+manager_mock
+manager_mock_services
+utils
